sysfs: fix sysfs_set_policy()
Fix a bug that causes a failure when trying to write a new cpufreq
policy whose maximum frequency value is less than the old policy's
minimum frequency value. Since the max value is not allowed to be
lower than the min value, writing to the sysfs file scaling_max_freq
will fail. The patch checks for this situation, and if it occurs, will
write the new min value before writing the new max value.

Signed-off-by: Joseph Xu <josephzxu@gmail.com>
Signed-off-by: Dominik Brodowski <linux@dominikbrodowski.net>